add check for file

+++ deploy_scripts/substitute_env.sh
@@ -5,6 +5,9 @@ for TEMPLATE in scorpio/config.py.deploy \
   appspec.yml.deploy \
   deploy_scripts/*.deploy
 do
-  envsubst < "$TEMPLATE" > echo "$TEMPLATE" | sed -e 's/\(\.deploy\)*$//g'
-  rm $TEMPLATE
+  if [[ -f "$TEMPLATE" ]]; then
+    echo "Replacing variables in $TEMPLATE"
+    envsubst < "$TEMPLATE" > `echo "$TEMPLATE" | sed -e 's/\(\.deploy\)*$//g'`
+    rm $TEMPLATE
+  fi
 done
